Rename gulp-deploy-app to gulp-deploy
